requirements.txt
cffi==1.16.0
colorama==0.4.6
cryptography==41.0.7
mysql-connector-python==8.2.0
numpy==1.26.2
oracledb==1.4.2
pandas==2.1.3
protobuf==4.21.12
pycparser==2.21
PyMySQL==1.1.0
pyodbc==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
